move filesystem utils to own file
